seperated out the main function from the sql_schema.py file
